+++ init-scripts/02_enable_timescaledb.sql
@@ -3,3 +3,6 @@ CREATE EXTENSION IF NOT EXISTS timescaledb;
 
 \c db_gateway
 CREATE EXTENSION IF NOT EXISTS timescaledb;
+
+\c db_market_data
+CREATE EXTENSION IF NOT EXISTS timescaledb;